--- src/main/java/com/miranda/chatop/configuration/JwtAuthFilter.java
@@ -1,44 +1,0 @@
-package com.miranda.chatop.configuration;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-//cette methode récuperère le header "Authorization de la requette HTTP"
-//utilise userAuthProvider pour valider  le jeton et configurer l'authentication.
-@RequiredArgsConstructor
-public class JwtAuthFilter extends OncePerRequestFilter {
-
-    private final UserAuthProvider userAuthProvider;
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                    HttpServletResponse response,
-                                    FilterChain filterChain) throws ServletException, IOException {
-        String header = request.getHeader(HttpHeaders.AUTHORIZATION);
-
-        if (header != null){
-            String[] authElements = header.split(" ");
-
-            if (authElements.length == 2 && "Bearer".equals(authElements[0])){
-                try {
-                    SecurityContextHolder.getContext().setAuthentication(userAuthProvider.validateToken(authElements[1]));
-                }catch (RuntimeException e){
-                    SecurityContextHolder.clearContext();
-                    throw e;
-                }
-             }
-       }
-
-        filterChain.doFilter(request,response);
-
-    }
-}
