--- src/main/java/com/miranda/chatop/configuration/RestExceptionHandler.java
@@ -1,20 +1,0 @@
-package com.miranda.chatop.configuration;
-
-import com.miranda.chatop.model.dtos.ErrorDto;
-import com.miranda.chatop.exception.AppException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@ControllerAdvice
-public class RestExceptionHandler {
-
-    @ExceptionHandler(value = {AppException.class})
-    @ResponseBody
-    public ResponseEntity<ErrorDto> handleException(AppException ex) {
-        return ResponseEntity.status(ex.getHttpStatus())
-                .body(new ErrorDto(ex.getMessage()));
-    }
-
-}
