--- src/main/java/com/miranda/chatop/configuration/SecurityConfig.java
@@ -1,49 +1,0 @@
-package com.miranda.chatop.configuration;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-import org.springframework.security.web.firewall.HttpFirewall;
-import org.springframework.security.web.firewall.StrictHttpFirewall;
-
-@RequiredArgsConstructor
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-    private final UserAuthProvider userAuthProvider;
-
-    //cette configuration Spring Security définit les règles de sécurité pour l'application.
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.csrf(AbstractHttpConfigurer::disable)
-                .addFilterBefore(new JwtAuthFilter(userAuthProvider), BasicAuthenticationFilter.class)
-                .sessionManagement(customizer -> customizer.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-
-                .authorizeHttpRequests((requests) ->
-                        requests
-                                .requestMatchers(HttpMethod.POST, "/auth/login","/auth/register", "/fileSystem").permitAll()
-                                .requestMatchers(HttpMethod.GET, "/swagger-ui/**", "/v3/api-docs/**", "/swagger-ui.html","/fileSystem/*").permitAll()
-                                //.requestMatchers(HttpMethod.PUT, "/rentals/*").permitAll()
-                                .anyRequest().authenticated()
-
-                );
-        return  http.build();
-    }
-
-    //le pare-feu HTTP strict pour contrôler les URL autorisées.
-    @Bean
-    public HttpFirewall strictHttpFirewall() {
-        StrictHttpFirewall firewall = new StrictHttpFirewall();
-        firewall.setAllowUrlEncodedDoubleSlash(true); // Autoriser les barres obliques encodées dans les URL
-        return firewall;
-    }
-}
-
