--- src/main/java/com/miranda/chatop/configuration/package-info.java
@@ -1,1 +1,0 @@
-package com.miranda.chatop.configuration;
