--- src/main/java/com/miranda/chatop/exception/AppException.java
@@ -1,16 +1,0 @@
-package com.miranda.chatop.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class AppException extends RuntimeException {
-    private final HttpStatus httpStatus;
-
-    public AppException(String message, HttpStatus httpStatus) {
-        super(message);
-        this.httpStatus = httpStatus;
-    }
-
-    public HttpStatus getHttpStatus(){
-        return httpStatus;
-    }
-}
