--- src/main/java/com/miranda/chatop/message/ResponseMessage.java
@@ -1,18 +1,0 @@
-package com.miranda.chatop.message;
-
-public class ResponseMessage {
-    private String message;
-
-    public ResponseMessage(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-}
