--- src/main/java/com/miranda/chatop/model/dtos/CreateRentalDto.java
@@ -1,2 +1,0 @@
-package com.miranda.chatop.model.dtos;
-
