--- src/main/java/com/miranda/chatop/model/dtos/CredentialsDto.java
@@ -1,4 +1,0 @@
-package com.miranda.chatop.model.dtos;
-//cette classe record est une classe immuable qui est  utilisée pour représenter des données simples et transférables.
-public record CredentialsDto (String email, String password){
-}
