--- src/main/java/com/miranda/chatop/model/dtos/ErrorDto.java
@@ -1,5 +1,0 @@
-package com.miranda.chatop.model.dtos;
-
-public record ErrorDto (String message){
-
-}
