--- src/main/java/com/miranda/chatop/repositories/MessageRepository.java
@@ -1,8 +1,0 @@
-package com.miranda.chatop.repositories;
-
-import com.miranda.chatop.model.entities.MessageEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
-    public MessageEntity findByMessage(String message);
-}
