--- src/main/java/com/miranda/chatop/repositories/UserRepository.java
@@ -1,11 +1,0 @@
-package com.miranda.chatop.repositories;
-
-import com.miranda.chatop.model.entities.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-    Optional<UserEntity> findByEmail(String email);
-}
-
