+++ src/main/java/com/miranda/chatop/repository/MessageRepository.java
@@ -1,0 +1,9 @@
+package com.miranda.chatop.repository;
+
+import com.miranda.chatop.model.entity.MessageEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface MessageRepository extends CrudRepository<MessageEntity, Long> {
+    public MessageEntity findByRental(Integer rental);
+
+}
