+++ src/main/java/com/miranda/chatop/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.miranda.chatop.repository;
+
+import com.miranda.chatop.model.entity.UserEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+    public UserEntity findByName(String name);
+}
+
