+++ src/main/java/com/miranda/chatop/service/IUserService.java
@@ -1,0 +1,5 @@
+package com.miranda.chatop.service;
+
+public interface IUserService {
+
+}
