--- src/main/java/com/miranda/chatop/services/IMessageService.java
@@ -1,7 +1,0 @@
-package com.miranda.chatop.services;
-
-import com.miranda.chatop.model.dtos.MessageDto;
-
-public interface IMessageService {
-    MessageDto saveMessageEntity(MessageDto messageDto);
-}
