+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
 server.servlet.contextPath=/api
-security.jwt.token.secret-key=my-secret-key
 
 spring.datasource.url=jdbc:mysql://localhost:3306/chatop
 spring.datasource.username=root
@@ -10,11 +9,3 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.generate.ddl=true
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-
-file.upload.dir=C:\\dev\\Projet3\\chatop\\uploads
-
-spring.servlet.multipart.enabled=true
-spring.servlet.multipart.max-file-size=10MB
-spring.servlet.multipart.max-request-size=10MB
-
-#springdoc.swagger-ui.path:/swagger-ui.html
